classes in plots are now integers if not strings

+++ adenine/utils/data_source.py
@@ -34,7 +34,7 @@ def generate_gauss(mu=(), std=(), n_sample=()):
     n_class, n_var = mu.shape
 
     X = np.zeros((n_sample*n_class, n_var))
-    y = np.zeros(n_sample*n_class)
+    y = np.zeros(n_sample*n_class, dtype=int)
 
     start = 0
     for i, s, m in zip(range(n_class), std, mu):
